feat(types): `hasKeyword`関数を`Bookmark.ts`に移動
`hasKeyword`関数はブックマークオブジェクトに関する純粋なヘルパー関数であり、特定のコンポーネントに依存するものではありません。
この関数を`Bookmark`型が定義されている`Bookmark.ts`に移動することで、再利用性が向上し、関心の分離が促進されます。

`BookmarkTable.tsx`は、移動された`hasKeyword`関数を`../types/Bookmark`からインポートするように修正しました。
